requirements.txt
ping3==5.1.5
python-dotenv==1.2.2
textbelt-py==0.1.6
